+++ app/test/db_test.py
@@ -1,4 +1,4 @@
-from . import db
+from .. import db
 
 
 def test_summer_looks():
